refactor: move HomePageTabImageNav
