--- src/components/Badge/index.ts
@@ -1,2 +1,0 @@
-export { default } from "./Badge";
-export type { BadgeProps, BadgeConfig } from "./Badge";
