+++ src/components/GenericPagesComponents/CurriculumTabBadge/CurriculumTabBadge.test.tsx
@@ -1,12 +1,12 @@
-import Badge from ".";
+import CurriculumTabBadge from ".";
 
 import theme from "@/styles/theme/default.theme";
 import renderWithTheme from "@/__tests__/__helpers__/renderWithTheme";
 
-describe("components/Badge", () => {
+describe("components/CurriculumTabBadge", () => {
   test("should be the correct size (from theme)", async () => {
     const { getByTestId } = renderWithTheme(
-      <Badge data-testid="test" text="Win" />,
+      <CurriculumTabBadge data-testid="test" text="Win" />,
     );
 
     expect(getByTestId("test")).toHaveStyle(`width: ${theme.badge.size}`);
