+++ src/components/GenericPagesComponents/CurriculumTabBadge/index.tsx
@@ -1,0 +1,5 @@
+export { default } from "./CurriculumTabBadge";
+export type {
+  CurriculumTabBadgeProps as BadgeProps,
+  BadgeConfig,
+} from "./CurriculumTabBadge";
