+++ src/components/GenericPagesComponents/HomePageTabImageNav/index.tsx
